remove pfbjobtest
